+++ src/main/java/org/kitteh/irc/client/library/element/ISupportParameter.java
@@ -24,6 +24,8 @@
 package org.kitteh.irc.client.library.element;
 
 import org.kitteh.irc.client.library.Client;
+import org.kitteh.irc.client.library.element.mode.ChannelMode;
+import org.kitteh.irc.client.library.element.mode.ChannelUserMode;
 
 import javax.annotation.Nonnull;
 import java.util.List;
