+++ src/main/java/org/kitteh/irc/client/library/element/mode/ChannelMode.java
@@ -21,7 +21,7 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.kitteh.irc.client.library.element;
+package org.kitteh.irc.client.library.element.mode;
 
 import org.kitteh.irc.client.library.Client;
