+++ src/main/java/org/kitteh/irc/client/library/event/channel/ChannelModeEvent.java
@@ -26,7 +26,7 @@ package org.kitteh.irc.client.library.event.channel;
 import org.kitteh.irc.client.library.Client;
 import org.kitteh.irc.client.library.element.Actor;
 import org.kitteh.irc.client.library.element.Channel;
-import org.kitteh.irc.client.library.element.ChannelModeStatusList;
+import org.kitteh.irc.client.library.element.mode.ChannelModeStatusList;
 import org.kitteh.irc.client.library.element.ServerMessage;
 import org.kitteh.irc.client.library.event.abstractbase.ActorChannelEventBase;
