+++ src/main/java/org/kitteh/irc/client/library/feature/ServerInfo.java
@@ -23,8 +23,8 @@
  */
 package org.kitteh.irc.client.library.feature;
 
-import org.kitteh.irc.client.library.element.ChannelMode;
-import org.kitteh.irc.client.library.element.ChannelUserMode;
+import org.kitteh.irc.client.library.element.mode.ChannelMode;
+import org.kitteh.irc.client.library.element.mode.ChannelUserMode;
 import org.kitteh.irc.client.library.element.ISupportParameter;
 
 import javax.annotation.Nonnull;
