+++ src/main/java/org/kitteh/irc/client/library/implementation/ActorProvider.java
@@ -27,10 +27,10 @@ package org.kitteh.irc.client.library.implementation;
 import org.kitteh.irc.client.library.Client;
 import org.kitteh.irc.client.library.element.Actor;
 import org.kitteh.irc.client.library.element.Channel;
-import org.kitteh.irc.client.library.element.ChannelMode;
-import org.kitteh.irc.client.library.element.ChannelModeStatus;
-import org.kitteh.irc.client.library.element.ChannelModeStatusList;
-import org.kitteh.irc.client.library.element.ChannelUserMode;
+import org.kitteh.irc.client.library.element.mode.ChannelMode;
+import org.kitteh.irc.client.library.element.mode.ChannelModeStatus;
+import org.kitteh.irc.client.library.element.mode.ChannelModeStatusList;
+import org.kitteh.irc.client.library.element.mode.ChannelUserMode;
 import org.kitteh.irc.client.library.element.ISupportParameter;
 import org.kitteh.irc.client.library.element.Staleable;
 import org.kitteh.irc.client.library.element.User;
